requirements.txt
boto3
numpy
opencv-python
PILLOW
botocore